+++ .eslintrc.js
@@ -27,6 +27,7 @@ module.exports = {
     "max-classes-per-file": ["error", 10],
     "import/prefer-default-export": "off",
     "object-curly-newline": "off",
+    "no-restricted-syntax": ["error", "ForInStatement", "LabeledStatement", "WithStatement"],
     "no-use-before-define": "off",
     "no-unused-vars": "off",
     "@typescript-eslint/no-use-before-define": ["error", { functions: false, classes: false }],